+++ packages/calcite-components/src/components.d.ts
@@ -4659,7 +4659,11 @@ export namespace Components {
           * Specifies the number of columns the component should span.
          */
         "colSpan": number;
+        /**
+          * When true, prevents user interaction.  Notes:  This prop should use the
+         */
         "disabled": boolean;
+        "lastCell": boolean;
         /**
           * Use this property to override individual strings used by the component.
          */
@@ -4703,6 +4707,7 @@ export namespace Components {
           * A heading to display above description content.
          */
         "heading": string;
+        "lastCell": boolean;
         /**
           * Use this property to override individual strings used by the component.
          */
@@ -4737,6 +4742,7 @@ export namespace Components {
           * When `true`, interaction is prevented and the component is displayed with lower opacity.
          */
         "disabled": boolean;
+        "lastVisibleRow": boolean;
         "numbered": boolean;
         "positionAll": number;
         "positionSection": number;
@@ -12087,7 +12093,11 @@ declare namespace LocalJSX {
           * Specifies the number of columns the component should span.
          */
         "colSpan"?: number;
+        /**
+          * When true, prevents user interaction.  Notes:  This prop should use the
+         */
         "disabled"?: boolean;
+        "lastCell"?: boolean;
         /**
           * Use this property to override individual strings used by the component.
          */
@@ -12127,6 +12137,7 @@ declare namespace LocalJSX {
           * A heading to display above description content.
          */
         "heading"?: string;
+        "lastCell"?: boolean;
         /**
           * Use this property to override individual strings used by the component.
          */
@@ -12157,6 +12168,7 @@ declare namespace LocalJSX {
           * When `true`, interaction is prevented and the component is displayed with lower opacity.
          */
         "disabled"?: boolean;
+        "lastVisibleRow"?: boolean;
         "numbered"?: boolean;
         "onCalciteInternalTableRowFocusRequest"?: (event: CalciteTableRowCustomEvent<TableRowFocusEvent>) => void;
         /**