+++ packages/calcite-components-angular/projects/component-library/CHANGELOG.md
@@ -3,17 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.0-next.4](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.3...@esri/calcite-components-angular@2.2.0-next.4) (2023-12-21)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
+## [2.2.0-next.3](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.2...@esri/calcite-components-angular@2.2.0-next.3) (2023-12-21)
+
+**Note:** Version bump only for package @esri/calcite-components-angular
+
 ## [2.2.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.1...@esri/calcite-components-angular@2.2.0-next.2) (2023-12-20)
 
-__Note:__ Version bump only for package @esri/calcite-components-angular
+**Note:** Version bump only for package @esri/calcite-components-angular
 
 ## [2.2.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.2.0-next.0...@esri/calcite-components-angular@2.2.0-next.1) (2023-12-20)
 
-__Note:__ Version bump only for package @esri/calcite-components-angular
+**Note:** Version bump only for package @esri/calcite-components-angular
 
 ## [2.2.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.1.0...@esri/calcite-components-angular@2.2.0-next.0) (2023-12-19)
 
-__Note:__ Version bump only for package @esri/calcite-components-angular
+**Note:** Version bump only for package @esri/calcite-components-angular
 
 ## [2.1.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-angular@2.0.0...@esri/calcite-components-angular@2.1.0) (2023-12-19)